// packages/editable-components/src/components/SectionWrapper/SectionWrapper.types.ts
import { Spacing } from "@easyblocks/core";

export type SectionWrapperCompiledValues = {
  containerMargin: Spacing;
  containerMaxWidth: string;
  escapeMargin: boolean;
  hide: boolean;
  headerMode: string;
  layout1Stack: string;
  layout2Stacks: string;
  layout2StacksVerticalAlign: string;
  headerStacksGap: Spacing;
  headerSectionGap: Spacing;
  footerSectionGap: Spacing;
  HeaderStack: any[];
  HeaderSecondaryStack: any[];
  Background__: any[];
  padding: Spacing;
  Component: any[];
};